requirements/common.txt
python-dotenv==1.1.0
bcrypt==4.3.0
flask
folium==0.19.5
flask_limiter==3.12.0
apscheduler==3.11.0
flask_socketio==5.5.1
Flask-WTF==1.2.2